gthread: Allow compiling g_once_init_enter for CHERI
We are actually loading a gsize which is not guaranteed to be the same as
a pointer (e.g. on CHERI-enabled systems). For such systems, we have to
explicitly use a size_t atomic operation since using a pointer-type on
would trigger a bounds errors. We could add g_atomic_size_exchange, but
as this is currently only needed for platforms with a modern LLVM compiler
we can use the builtin directly.

(cherry picked from commit 56ac48032287ad3d7d65de1c6377f949f8732bf8)